Rename config.rs to settings.rs
