+++ verification/src/lib.rs
@@ -1,15 +1,15 @@
 mod compiler;
-mod config;
 mod consts;
 mod http_server;
 mod scheduler;
+mod settings;
 mod solidity;
 mod types;
 
 #[cfg(test)]
 mod tests;
 
-pub use self::config::Settings;
+pub use self::settings::Settings;
 pub use ethers_core::types::Bytes as DisplayBytes;
 pub use http_server::{
     configure_router,
